src/main/java/cn/mark/o2o/service/ShopService.java
package cn.mark.o2o.service;


import cn.mark.o2o.dto.ImageHolder;
import cn.mark.o2o.dto.ShopExecution;
import cn.mark.o2o.entity.Shop;
import cn.mark.o2o.exceptions.ShopOperationException;

public interface ShopService {
	/**
	 * 根据shopCondition分页返回店铺列表
	 * @param shopConditon
	 * @param pageIndex
	 * @param pageSize
	 * @return
	 */
	public ShopExecution getShopList(Shop shopCondition,int pageIndex,int pageSize);
	
	/**
	 * 通过店铺Id获取店铺信息
	 * @param shopId
	 * @return
	 */
	Shop getByShopId(long shopId);
	
	/**
	 * 更新店铺信息，包括图片处理
	 * @param shop
	 * @param shopImgInputStream
	 * @param fileName
	 * @return
	 */
	ShopExecution modifyShop(Shop shop, ImageHolder thumbnail) throws ShopOperationException;
	
	/**
	 * 注册店铺信息，包括图片处理
	 * @param shop
	 * @param shopImgInputstream
	 * @param fileName
	 * @return
	 * @throws ShopOperationException
	 */
	ShopExecution addShop(Shop shop,ImageHolder thumbnail) throws ShopOperationException;
}
